+++ Mods/S7_Config_de8f15f2-65a2-4ee4-a25f-7a7ab0305a58/Story/RawFiles/Lua/BootstrapClient.lua
@@ -12,11 +12,11 @@ logSource = "Lua:BootstrapClient"
 --  STATS LOADER
 --  ============
 
-local function S7_StatsLoader() --  Loads stats-configuration json after StatsLoaded Event.
-    if S7_ConfigSettings.StatsLoader.Enable == true then --  StatsLoader enabled in settings.
-        S7_ConfigLog("Loading " .. S7_ConfigSettings.StatsLoader.FileName)
-        local file = Ext.LoadFile(S7_ConfigSettings.StatsLoader.FileName) or "" --  Load file if it exists. Load empty string otherwise.
-        if file ~= nil and file ~= "" then --  if configData file exists and is not empty.
+local function StatsLoader() --  Loads stats-configuration json after StatsLoaded Event.
+    if ConfigSettings.StatsLoader.Enable == true then --  StatsLoader enabled in settings.
+        S7_ConfigLog("Loading " .. ConfigSettings.StatsLoader.FileName)
+        local file = Ext.LoadFile(ConfigSettings.StatsLoader.FileName) or "" --  Load file if it exists. Load empty string otherwise.
+        if S7_ValidJSON(file) then --  if configData file exists and is not empty.
             local configData = Ext.JsonParse(file) --  Parse into table.
             for modID, content in pairs(configData) do --  for each mod's configData
                 table.insert(toConfigure, {[modID] = content}) --  Queue files for configuration.
@@ -26,12 +26,14 @@ local function S7_StatsLoader() --  Loads stats-configuration json after StatsLo
         toConfigure = {} -- flush config queue
         toSync = {} --  flush Sync queue
         S7_ConfigLog("StatsLoading Completed.")
+    else --  if StatsLoader disabled in settings.
+        S7_ConfigLog("StatsLoader is disabled.", "[Warning]")
     end
 end
 
---  ===============================================
-Ext.RegisterListener("StatsLoaded", S7_StatsLoader)
---  ===============================================
+--  ============================================
+Ext.RegisterListener("StatsLoaded", StatsLoader)
+--  ============================================
 
 --  ===============
 --  CATCH BROADCAST
@@ -39,12 +41,12 @@ Ext.RegisterListener("StatsLoaded", S7_StatsLoader)
 
 local function S7_CatchBroadcast(channel, payload) --  Listens for broadcasts from the Server.
     if channel == "S7_ConfigData" then --  if broadcast channel is S7_ConfigData.
-        S7_ConfigLog("Client recieved Active Configuration. Saving File: " .. S7_ConfigSettings.StatsLoader.FileName)
-        Ext.SaveFile(S7_ConfigSettings.StatsLoader.FileName, payload) --  Save stringified json.
+        S7_ConfigLog("Client recieved Active Configuration. Saving File: " .. ConfigSettings.StatsLoader.FileName)
+        Ext.SaveFile(ConfigSettings.StatsLoader.FileName, payload) --  Save stringified json.
     end
 
     if channel == "S7_ValidateClientConfig" then --  if broadcast channel is S7_ValidateClientConfig
-        local verify = Ext.LoadFile(S7_ConfigSettings.StatsLoader.FileName) or "" --    Load local ActiveConfiguration if available
+        local verify = Ext.LoadFile(ConfigSettings.StatsLoader.FileName) or "" --    Load local ActiveConfiguration if available
         if payload == verify then
             Ext.PostMessageToServer("S7_ValidateClientResponse", "All Good.")
         else