+++ Mods/S7_Config_de8f15f2-65a2-4ee4-a25f-7a7ab0305a58/Story/RawFiles/Lua/BootstrapServer.lua
@@ -18,13 +18,13 @@ local function S7_Config_ModMenuRelay(Signal) --  Signal recieved from Osiris.
     -- ====================
 
     if Signal == "S7_StatsConfigurator" then
-        local file = Ext.LoadFile(S7_ConfigSettings.ConfigFile) or "" --  Load file.
+        local file = Ext.LoadFile(ConfigSettings.ConfigFile) or "" --  Load file.
         if type(file) == "string" and file ~= nil and file ~= "" then --  if file exists and is not empty.
-            S7_ConfigLog("Loading: " .. S7_ConfigSettings.ConfigFile)
+            S7_ConfigLog("Loading: " .. ConfigSettings.ConfigFile)
             table.insert(toConfigure, {["S7_Config"] = file}) -- Queue json for Configuration.
         else
-            S7_ConfigLog(S7_ConfigSettings.ConfigFile .. " not found. Creating empty file.", "[Error]")
-            Ext.SaveFile(S7_ConfigSettings.ConfigFile, "")
+            S7_ConfigLog(ConfigSettings.ConfigFile .. " not found. Creating empty file.", "[Error]")
+            Ext.SaveFile(ConfigSettings.ConfigFile, "")
         end
         S7_StatsConfigurator() --  Calls StatsConfigurator.
         S7_StatsSynchronize() --  Synchronize stats for all clients.
@@ -37,8 +37,8 @@ local function S7_Config_ModMenuRelay(Signal) --  Signal recieved from Osiris.
 
     if Signal == "S7_StatsSynchronize" then
         S7_ConfigLog("Synchronizing stats at Player's request.")
-        if S7_ConfigSettings.ManuallySynchronize ~= nil then --  Checks if player wants to manually synchronize certain stats.
-            for i, stats in ipairs(S7_ConfigSettings.ManuallySynchronize) do --  Iterate over manually selected stats.
+        if ConfigSettings.ManuallySynchronize ~= nil then --  Checks if player wants to manually synchronize certain stats.
+            for i, stats in ipairs(ConfigSettings.ManuallySynchronize) do --  Iterate over manually selected stats.
                 table.insert(toSync, stats) --  insert stats into toSync queue.
             end
         end
@@ -50,16 +50,16 @@ local function S7_Config_ModMenuRelay(Signal) --  Signal recieved from Osiris.
     --  =================
 
     if Signal == "S7_BuildConfigData" then
-        local buildData = Ext.LoadFile(S7_ConfigSettings.ConfigFile) or ""
+        local buildData = Ext.LoadFile(ConfigSettings.ConfigFile) or ""
         S7_BuildConfigData("S7_Config", buildData)
-        S7_ConfigLog("Rebuilt " .. S7_ConfigSettings.StatsLoader.FileName .. " using " .. S7_ConfigSettings.ConfigFile)
+        S7_ConfigLog("Rebuilt " .. ConfigSettings.StatsLoader.FileName .. " using " .. ConfigSettings.ConfigFile)
     end
 
     --  SEND CONFIG DATA
     --  ================
 
     if Signal == "S7_BroadcastConfigData" then
-        local broadcast = Ext.LoadFile(S7_ConfigSettings.StatsLoader.FileName) or ""
+        local broadcast = Ext.LoadFile(ConfigSettings.StatsLoader.FileName) or ""
         if type(broadcast) == "string" and broadcast ~= "" and broadcast ~= nil then --  if file exists and is not empty
             Ext.BroadcastMessage("S7_ConfigData", broadcast) --  broadcast Server's configFile
             S7_ConfigLog("Server broadcasts Active Configuration Profile.")
@@ -72,13 +72,13 @@ local function S7_Config_ModMenuRelay(Signal) --  Signal recieved from Osiris.
     --  =====================
 
     if Signal == "S7_ValidateClientConfig" then
-        local compare = Ext.LoadFile(S7_ConfigSettings.StatsLoader.FileName) --  Loads server's config
+        local compare = Ext.LoadFile(ConfigSettings.StatsLoader.FileName) --  Loads server's config
         if type(compare) == "string" and compare ~= "" and compare ~= nil then -- if file exists and is not empty
             Ext.BroadcastMessage("S7_ValidateClientConfig", compare) -- broadcast server's config file to all clients.
             S7_ConfigLog("Validating Client Config.")
         else
             S7_ConfigLog(
-                "Nothing to validate. Please check if the server has " .. S7_ConfigSettings.StatsLoader.FileName,
+                "Nothing to validate. Please check if the server has " .. ConfigSettings.StatsLoader.FileName,
                 "[Error]"
             )
         end
@@ -88,11 +88,11 @@ local function S7_Config_ModMenuRelay(Signal) --  Signal recieved from Osiris.
     --  ==================
 
     if Signal == "S7_ToggleStatsLoader" then
-        if S7_ConfigSettings.StatsLoader.Enable == true then
-            S7_ConfigSettings.StatsLoader.Enable = false
+        if ConfigSettings.StatsLoader.Enable == true then
+            ConfigSettings.StatsLoader.Enable = false
             S7_ConfigLog("StatsLoader: Deactivated")
         else
-            S7_ConfigSettings.StatsLoader.Enable = true
+            ConfigSettings.StatsLoader.Enable = true
             S7_ConfigLog("StatsLoader: Activated")
         end
     end
@@ -101,11 +101,11 @@ local function S7_Config_ModMenuRelay(Signal) --  Signal recieved from Osiris.
     --  ============================
 
     if Signal == "S7_ToggleSyncStatPersistence" then
-        if S7_ConfigSettings.SyncStatPersistence == true then
-            S7_ConfigSettings.SyncStatPersistence = false
+        if ConfigSettings.SyncStatPersistence == true then
+            ConfigSettings.SyncStatPersistence = false
             S7_ConfigLog("SyncStatPersistence: Deactivated")
         else
-            S7_ConfigSettings.SyncStatPersistence = true
+            ConfigSettings.SyncStatPersistence = true
             S7_ConfigLog("SyncStatPersistence: Activated")
         end
     end
@@ -114,11 +114,11 @@ local function S7_Config_ModMenuRelay(Signal) --  Signal recieved from Osiris.
     --  ==========================
 
     if Signal == "S7_ToggleSafetyCheck" then
-        if S7_ConfigSettings.BypassSafetyCheck == false then
-            S7_ConfigSettings.BypassSafetyCheck = true
+        if ConfigSettings.BypassSafetyCheck == false then
+            ConfigSettings.BypassSafetyCheck = true
             S7_ConfigLog("BypassSafetyCheck: Activated")
         else
-            S7_ConfigSettings.BypassSafetyCheck = false
+            ConfigSettings.BypassSafetyCheck = false
             S7_ConfigLog("BypassSafetyCheck: Deactivated")
         end
     end
@@ -127,7 +127,7 @@ local function S7_Config_ModMenuRelay(Signal) --  Signal recieved from Osiris.
     -- ======================
 
     if Signal == "S7_SetDefaultSettings" then --  Resets ConfigSettings to Default Values.
-        S7_ConfigSettings = S7_Rematerialize(S7_DefaultSettings)
+        ConfigSettings = S7_Rematerialize(DefaultSettings)
         S7_ConfigLog("Using default settings.")
     end
 
@@ -142,9 +142,9 @@ local function S7_Config_ModMenuRelay(Signal) --  Signal recieved from Osiris.
     --  EXPORT CURRENT SETTINGS
     -- =========================
 
-    if Signal == "S7_ExportCurrentSettings" then --  Exports the current ConfigSettings to S7_ConfigSettings.json file.
-        local exportSettings = Ext.JsonStringify(S7_ConfigSettings) --  stringifies the current ConfigSettings.
-        Ext.SaveFile("S7_ConfigSettings.json", exportSettings) --  Save json file.
+    if Signal == "S7_ExportCurrentSettings" then --  Exports the current ConfigSettings to ConfigSettings.json file.
+        local exportSettings = Ext.JsonStringify(ConfigSettings) --  stringifies the current ConfigSettings.
+        Ext.SaveFile("ConfigSettings.json", exportSettings) --  Save json file.
         S7_ConfigLog("Exporting current ConfigSettings to S7_ConfigSettings.json")
         S7_RefreshSettings() --  Reload settings.
         S7_ConfigLog("Custom Settings Exported and Refreshed.")
@@ -155,7 +155,7 @@ local function S7_Config_ModMenuRelay(Signal) --  Signal recieved from Osiris.
 
     if Signal == "S7_StatsExportTSV" then
         S7_StatsExportTSV() --  Logs statIDs in an external TSV file for reference
-        S7_ConfigLog("Exporting StatIDs to " .. S7_ConfigSettings.ExportStatIDtoTSV.FileName)
+        S7_ConfigLog("Exporting StatIDs to " .. ConfigSettings.ExportStatIDtoTSV.FileName)
     end
 
     --  CHANGELOG
@@ -200,11 +200,11 @@ local function S7_Config_ModMenuRelay(Signal) --  Signal recieved from Osiris.
     --  ==========
 
     if Signal == "S7_ToggleConfigLog" then
-        if S7_ConfigSettings.ConfigLog.Enable == true then
-            S7_ConfigSettings.ConfigLog.Enable = false
+        if ConfigSettings.ConfigLog.Enable == true then
+            ConfigSettings.ConfigLog.Enable = false
             S7_ConfigLog("S7_ConfigLog: Disabled", "[Warning]")
         else
-            S7_ConfigSettings.ConfigLog.Enable = true
+            ConfigSettings.ConfigLog.Enable = true
             S7_ConfigLog("S7_ConfigLog: Enabled", "[Warning]")
         end
     end