+++ Mods/S7_Config_de8f15f2-65a2-4ee4-a25f-7a7ab0305a58/Story/RawFiles/Lua/S7_ConfigCollections.lua
@@ -87,8 +87,8 @@ function S7_RebuildCollections()
     end
     --  CUSTOM-COLLECTIONS
     --  ==================
-    if S7_ConfigSettings.CustomCollections ~= nil then
-        for key, value in pairs(S7_ConfigSettings.CustomCollections) do
+    if ConfigSettings.CustomCollections ~= nil then
+        for key, value in pairs(ConfigSettings.CustomCollections) do
             configCollections[key] = S7_Rematerialize(value)
         end
     end